Adduct formation rules

Formation rules for electrospray ionisation adducts

Usage

adduct_rules()

adduct_names()

Format

A tibble containing 58 rows and 6 columns

Name

Adduct name

Charge

Charge added

xM

Number of M needed

Add

Mass value change

Nelec

Electron change

AddAt

pseudo formula of the atoms to be added to the molecular formula of one M

RemAt

pseudo formula of the atoms to be removed to the molecular formula of one M

AddEx

pseudo formula of the atoms to be added to the final ionisation product molecular formula

RemEx

pseudo formula of the atoms to be removed to the final ionisation product molecular formula

Rule

Structural rule for formation

Examples

adduct_names()
#>  [1] "[M+3Na]3+"          "[M+H+2Na]3+"        "[M+2H+Na]3+"       
#>  [4] "[M+3H]3+"           "[2M+3H2O+2H]2+"     "[M+3ACN+2H]2+"     
#>  [7] "[M+2ACN+2H]2+"      "[M+2Na]2+"          "[M+ACN+2H]2+"      
#> [10] "[M+H+K]2+"          "[M+H+Na]2+"         "[M+H+NH4]2+"       
#> [13] "[M+2H]2+"           "[M+2H-H2O-NH3]2+"   "[2M+ACN+Na]1+"     
#> [16] "[2M+ACN+H]1+"       "[2M+K]1+"           "[2M+K41]1+"        
#> [19] "[2M+Na]1+"          "[2M+NH4]1+"         "[2M+H]1+"          
#> [22] "[M+IsoProp+Na+H]1+" "[M+2ACN+H]1+"       "[M+DMSO+H]1+"      
#> [25] "[M+2K-H]1+"         "[M+ACN+Na]1+"       "[M+IsoProp+H]1+"   
#> [28] "[M+2Na-H]1+"        "[M+ACN+H]1+"        "[M+K]1+"           
#> [31] "[M+K41]1+"          "[M+H+CH3OH]1+"      "[M+Na]1+"          
#> [34] "[M+NH4]1+"          "[M+H]1+"            "[M1+.]1+"          
#> [37] "[M+H-NH3]1+"        "[M+H-H2O]1+"        "[M+H-FA]1+"        
#> [40] "M"                  "[3M-H]1-"           "[2M+Hac-H]1-"      
#> [43] "[2M+FA-H]1-"        "[2M+Na-2H]1-"       "[2M-H]1-"          
#> [46] "[M+TFA-H]1-"        "[M+Br]1-"           "[M+Hac-H]1-"       
#> [49] "[M+FA-H]1-"         "[M+K-2H]1-"         "[M+K41-2H]1-"      
#> [52] "[M+Cl]1-"           "[M+Cl37]1-"         "[M+Na-2H]1-"       
#> [55] "[M1-.]1-"           "[M-H]1-"            "[M-2H]2-"          
#> [58] "[M-3H]3-"          

adduct_rules()
#> # A tibble: 58 × 12
#>    Name   Charge    xM   Add Nelec AddAt RemAt AddEx RemEx Rule  Default Isoto…¹
#>    <chr>   <int> <int> <dbl> <int> <chr> <chr> <chr> <chr> <chr>   <int>   <dbl>
#>  1 [M+3N…      3     1 23.0     -3 NA    NA    Na3   NA    HBA2…       0       0
#>  2 [M+H+…      3     1 15.7     -3 NA    NA    Na2H  NA    HBA2…       0       0
#>  3 [M+2H…      3     1  8.33    -3 NA    NA    NaH2  NA    HBA2…       0       0
#>  4 [M+3H…      3     1  1.01    -3 NA    NA    H3    NA    HBA2…       0       0
#>  5 [2M+3…      2     2 28.0     -2 NA    NA    H8O3  NA    HBA2…       0       0
#>  6 [M+3A…      2     1 62.5     -2 NA    NA    C6H1… NA    HBA2…       0       0
#>  7 [M+2A…      2     1 42.0     -2 NA    NA    C4H8… NA    HBA2…       0       0
#>  8 [M+2N…      2     1 23.0     -2 NA    NA    Na2   NA    HBA2…       1       0
#>  9 [M+AC…      2     1 21.5     -2 NA    NA    C2H5N NA    HBA2…       0       0
#> 10 [M+H+…      2     1 20.0     -2 NA    NA    KH    NA    HBA2…       1       0
#> # … with 48 more rows, and abbreviated variable name ¹​Isotopic